% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlmail.R
\name{initiate_htmlmail}
\alias{initiate_htmlmail}
\title{Create an htmlmail instance}
\usage{
initiate_htmlmail(htmlpath)
}
\arguments{
\item{htmlpath}{the path of the html content for the email, especially created from mjml app.}
}
\value{
An environment, with create_draft(to, from, subject) method, and send(to, from, subject) method using gmail
}
\description{
must first run gm_auth_configure(path = "support/gmailAPI.json") with respect to the json file from gmail API credential setup (see gmailr).
}
\examples{
none
}
